+++ NEWS.md
@@ -1,4 +1,9 @@
 # TaxSEA News
 
+## Version 0.99.2 (2025-02-20)
+- Added ability to test custom taxon sets using `custom_db` parameter.
+- Removed deprecated `plotting.R` file.
+- TaxSEA now reports the KS test statistic for each taxon set.
+
 ## Version 0.99.1 (2025-02-11)
 - Bringing into line with requirements for BioConductor submission